Update todolist with mongoDB and create custom-made list for users

+++ index.js
@@ -1,10 +1,10 @@
 import express from "express";
 import mongoose from "mongoose";
-import e from "express";
+import _ from "lodash";
 
 const app = express();
 const port = 3000;
-const toDoList = [];
+
 function getDate() {
     const daysOfWeek = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
     const months = [
@@ -52,11 +52,13 @@ const item3 = new Item({
 
 const defaultItems = [item1, item2, item3];
 
-/*Item.insertMany(defaultItems)
-    .then(function () {
-        console.log("Successfully saved default items to todolistDB");
-    })
-    .catch((err)=>console.log(err));*/
+const listSchema = new mongoose.Schema({
+    name: String,
+    items: [itemSchema]
+});
+
+const List = mongoose.model("List", listSchema);
+
 
 app.get("/", (req, res) => {
     Item.find({})
@@ -70,27 +72,70 @@ app.get("/", (req, res) => {
                 // after insert items, redirect to the else branch
                 res.redirect("/");
             } else {
-                res.render("index.ejs", {list: foundItems, date: getDate(),})}})
+                res.render("index.ejs", {list: foundItems, title: getDate()})}})
         .catch((err) => console.log(err));
 
 });
 
+app.get("/:customListName", (req, res) => {
+    const customListName = _.capitalize(req.params.customListName);
+    List.findOne({name: customListName})
+        .then((foundList) => {
+            if (foundList) {
+                // show an existing list
+                res.render("index.ejs" , {list: foundList.items, title: foundList.name})
+            } else {
+                // create a new list
+                const list = new List({
+                    name: customListName,
+                    items: defaultItems
+                })
+                list.save();
+                res.redirect("/" + customListName);
+            }
+        })
+        .catch((err) => console.log(err))
+})
+
+
 app.post("/submit", (req, res) => {
-    const itemName = req.body["content"];
+    const itemName = req.body.content;
+    const listName = req.body.list;
+
     const newItem = new Item({
         name: itemName,
     });
-    newItem.save();
 
-    res.redirect("/");
+    if (listName === getDate()) {
+        newItem.save();
+        res.redirect("/");
+    } else {
+        List.findOne({name: listName})
+            .then((foundList) => {
+                // push the newItem to the items array
+                foundList.items.push(newItem);
+                // save the updated foundlist
+                foundList.save();
+                res.redirect("/" + listName);
+            })
+    }
 })
 
 app.post("/delete", (req, res) => {
     const checkItemId = req.body.checkbox;
-    Item.findOneAndDelete({_id: checkItemId})
-        .then((deletedItem) => console.log(`${deletedItem.name} is deleted.`))
-        .catch((err) => console.log(err))
-        .finally(() => res.redirect("/"))
+    const listName = req.body.listName;
+
+    if (listName === getDate()) {
+        Item.findOneAndDelete({_id: checkItemId})
+            .then((deletedItem) => console.log(`${deletedItem.name} is deleted.`))
+            .catch((err) => console.log(err))
+            .finally(() => res.redirect("/"))
+    } else {
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkItemId}}})
+            .then((foundList) => res.redirect("/" + listName))
+            .catch((err) => console.log(err))
+    };
+
 });
 
 app.listen(port, () => {
